requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
braceexpand==0.1.7
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.1
decorator==5.2.1
editdistance==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.11
inflect==7.5.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
nemo-toolkit==1.20.0
networkx==3.4.2
numba==0.62.1
numpy==1.24.3
omegaconf==2.3.0
onnx==1.19.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.8.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.31.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webdataset==1.0.2
Werkzeug==3.1.3
wget==3.2
wrapt==2.0.1
yarl==1.22.0
youtokentome==1.0.6